GJ-Kuijsten Software-Engineering-for-Cybersecurity feature-Frontend-UI src/backend/requirements.txt
fastapi
uvicorn[standard]
passlib[bcrypt]
python-jose[cryptography]
pydantic
jinja2
python-multipart
httpx
boto3
python-dotenv
